Scrapers/mathacademyscraper/requirements.txt
playwright>=1.45.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
supabase>=2.0.0
postgrest>=0.10.0